+++ Pot.java
@@ -1,7 +1,7 @@
 public class Pot{
 
     private static double potTotal = 0;
-    private static int highestBet = 0;
+    private static double highestBet = 0;
 
     public static void addToPot(double bet){                   // add to the pot total.
         Pot.potTotal += bet;
@@ -11,7 +11,7 @@ public class Pot{
         return potTotal;
     }
 
-    public static int getHighestBet(){                      // Return the current highest best
+    public static double getHighestBet(){                      // Return the current highest best
         return highestBet;
     }
 
@@ -19,7 +19,7 @@ public class Pot{
 
         potTotal = 0;
     }
-    public static void setHighestBet(int bet){
+    public static void setHighestBet(double bet){
         highestBet = bet;
     }
